rename hamming.rs to scalar.rs
