+++ src/scalar.rs
